JAVA-2754 - adapt mapper to SPI
