+++ pom.xml
@@ -36,7 +36,7 @@
     <maven-surefire-plugin.version>3.0.0-M4</maven-surefire-plugin.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <datastax-java-driver.version>4.7.0-java2792+20200602064812</datastax-java-driver.version>
+    <datastax-java-driver.version>4.7.0-java2792-SNAPSHO​T</datastax-java-driver.version>
     <!--suppress UnresolvedMavenProperty -->
     <graalvmHome>${env.GRAALVM_HOME}</graalvmHome>
   </properties>
