+++ runtime/src/main/resources/META-INF/services/com.datastax.oss.driver.api.mapper.result.MapperResultProducerService
@@ -1,0 +1,1 @@
+com.datastax.oss.quarkus.runtime.api.reactive.mapper.MutinyProducerService
